refactor(archtest): errcode 族迁 module-path-agnostic CellRule + 缩 baseline [#1302]
把 errcode 族 8 规则的检测逻辑 + 共享 helper 从 errcode_invariants_test.go
MOVE 到非 test errcode_invariants.go，platform 符号路径经 PlatformModulePath
派生（不再 bare "github.com/ghbvf/gocell" 字面量），_test.go 变薄 dogfood 调
同一 Check*（单源）。6 个外部可移植规则进 StandardCellRules()
（KIND-LITERAL / MESSAGE-CONST-LITERAL / ERROR-FIRST-API / ERROR-FIRST-TYPED-NIL
/ EXPORTED-ERROR-NEW / DETAILS-SEALED-FIELD-FROZEN）；PREFIX-OWNERSHIP /
CARVEOUT-ADR-CONSISTENCY 绑 gocell 自身 registry/ADR，外部 vacuous，迁形态但
不注册。从 module_path_funnel.baseline 删 errcode_invariants_test.go（ratchet-down）。

附带：promote MESSAGE-CONST-LITERAL-01 进 StandardCellRules 后，
adapters/postgres/migration_set.go:106 的 errcode.Wrap fmt.Sprintf message
（pre-existing 违规，develop nightly 已红）使 TestRunStandardCellRules 红 —
按 const-literal message + WithDetails(PublicString) 范式修正（与上下相邻
errcode.New 调用一致），闭环 gocell 自身满足所 promote 的规则。
