+++ adapters/postgres/migration_set.go
@@ -104,7 +104,8 @@ func (s *MigrationSet) addNamespace(ns migration.Namespace, fsys fs.FS) error {
 	// a malformed external-cell migration never silently vanishes at apply/verify.
 	if _, err := ExpectedVersion(fsys); err != nil {
 		return errcode.Wrap(errcode.KindInvalid, ErrAdapterPGMigrate,
-			fmt.Sprintf("postgres: migration set namespace %q has a malformed migration file", ns), err)
+			"postgres: migration set namespace has a malformed migration file", err,
+			errcode.WithDetails(errcode.PublicString("namespace", ns.String())))
 	}
 	for _, e := range s.entries {
 		if e.ns == ns {
